fix: preference save/load, menu icons, text rendering, and file removal
- Guard PRFN resource expansion to only apply to compact (default)
  records, preventing memmove from corrupting already-saved preferences
- Implement OpenFilePref/EraseFilePref with Pascal/C string detection;
  add mac_remove() for host-path file deletion
- Read castonfriends from preferences dialog item 30
- Add cicn icons to Scenario menu items (note keeper, journal)
- Support bitmap icons in WinMenuController popup menus with HBitmap cache
- Fix WinMenuSync to propagate enabled/disabled state correctly
- Fix SetItemIcon: cast iconIndex to uint8_t before adding 256
- Fix text wrap width (remove extra 50px) and skip vertical centering
  offset for multi-line text
- Remove double P→C conversion for journal strings in HandleMenuChoice
- Disable Scenario menu when charnum < 0

+++ src/MenuManager.cpp
@@ -263,7 +263,7 @@ void SetItemIcon(MenuHandle theMenu, int16_t item, int16_t iconIndex) {
   auto& menu_item = menu->items.at(item - 1);
   menu_item.icon_number = static_cast<uint8_t>(iconIndex);
   if (iconIndex != 0) {
-    int16_t cicn_id = iconIndex + 256;
+    int16_t cicn_id = static_cast<uint8_t>(iconIndex) + 256;
     auto handle = GetResource(ResourceDASM::RESOURCE_TYPE_cicn, cicn_id);
     if (handle) {
       auto cicn = ResourceFile::decode_cicn(*handle, GetHandleSize(handle));
